Adjust weight for A330-200 plus fuel capacity for P&W version
The -200 had an empty weight set to 214000 lbs which is about 50000
lighter than any source I found. Real life number varies between
sources and with engine versions, but this is a closer  match to
what SimBrief uses.

Note that the -300 is still set 214000 lbs and needs to be updated
as well but I haven't been flying that version.

Fuel capacity was around 30000 lbs under Airbus spec. Updated to
make wing and center tanks closer to correct. This will still be a
bit under spec since the trim tanks aren't implemented. Note this
was only done for the P&W version since the change is in an engine
specific version and that's the only one I have been flying.