Set dimming knobs on "load panel state" to make controls visible
Main panel knob, FCU panel knob, and overhead panel knob were left
set to 0 when loading panel states from the aircraft config dialog.
This made it difficult to see the controls (including the knobs) at
night. Now the knobs will be set to a non-zero value for the states
"Ready to Start Eng", "Ready for Taxi", and "Ready for takeoff".