Improve lower ECAM Flight Control page performance
Made some simple improvements to avoid setColor calls unless a status
has actually changed. This reduces page update time by about 70%.

Note that the page shows 5 spoilers per wing but the A330 actually
has 6. This is probably some A320 leftovers.

During fault testing it was noticed that a leak in the green
hydraulic system causes the pressure in the blue system to increase,
apparently without limit. This is not an ECAM issue but a problem
somewhere in the hydraulic code.